refactor: pull ImportedProduct into types
